Go back and just reuse `withFreePort`
