+++ hydra-node/test/Hydra/HeadLogicSpec.hs
@@ -55,7 +55,6 @@ spec :: Spec
 spec = do
   parallel $ do
     describe "Types" $ do
-      roundtripAndGoldenSpecs (Proxy @(Message SimpleTx))
       roundtripAndGoldenSpecs (Proxy @(Event SimpleTx))
       roundtripAndGoldenSpecs (Proxy @(HeadState SimpleTx))
