+++ hydra-node/test/Hydra/NetworkSpec.hs
@@ -2,7 +2,7 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 -- | Test the real networking layer
-module Hydra.Port.NetworkSpec (spec) where
+module Hydra.NetworkSpec where
 
 import Hydra.Prelude
 import Test.Hydra.Prelude
@@ -15,38 +15,42 @@ import Hydra.Logging (showLogsOnFailure)
 import Hydra.Network (Host (..), Network)
 import Hydra.Network.Message (Message (..))
 import Hydra.Network.Ouroboros (broadcast, withOuroborosNetwork)
+import Test.Aeson.GenericSpecs (roundtripAndGoldenSpecs)
 import Test.Network.Ports (randomUnusedTCPPorts)
-import Test.QuickCheck (property)
+import Test.QuickCheck (
+  property,
+ )
 import Test.QuickCheck.Instances.ByteString ()
 
-spec :: SpecWith (TVar IO [Int])
+spec :: Spec
 spec = do
   let lo = "127.0.0.1"
-  describe "Ouroboros Network" $
-    parallel $ do
-      it "broadcasts messages to single connected peer" $ \tvar -> do
-        received <- atomically newTQueue
-        showLogsOnFailure $ \tracer -> failAfter 30 $ do
-          [port1, port2] <- fmap fromIntegral <$> randomUnusedTCPPorts tvar 2
-          withOuroborosNetwork tracer (Host lo port1) [Host lo port2] (const @_ @Integer $ pure ()) $ \hn1 ->
-            withOuroborosNetwork @Integer tracer (Host lo port2) [Host lo port1] (atomically . writeTQueue received) $ \_ -> do
-              withNodeBroadcastingForever hn1 1 $
-                atomically (readTQueue received) `shouldReturn` 1
 
-      it "broadcasts messages between 3 connected peers" $ \tvar -> do
-        node1received <- atomically newTQueue
-        node2received <- atomically newTQueue
-        node3received <- atomically newTQueue
-        showLogsOnFailure $ \tracer -> failAfter 30 $ do
-          [port1, port2, port3] <- fmap fromIntegral <$> randomUnusedTCPPorts tvar 3
-          withOuroborosNetwork @Integer tracer (Host lo port1) [Host lo port2, Host lo port3] (atomically . writeTQueue node1received) $ \hn1 ->
-            withOuroborosNetwork tracer (Host lo port2) [Host lo port1, Host lo port3] (atomically . writeTQueue node2received) $ \hn2 -> do
-              withOuroborosNetwork tracer (Host lo port3) [Host lo port1, Host lo port2] (atomically . writeTQueue node3received) $ \hn3 -> do
-                withNodesBroadcastingForever [(hn1, 1), (hn2, 2), (hn3, 3)] $
-                  assertAllnodesReceivedMessagesFromAllOtherNodes [(node1received, 1), (node2received, 2), (node3received, 3)]
+  describe "Ouroboros Network" $ do
+    it "broadcasts messages to single connected peer" $ do
+      received <- atomically newTQueue
+      showLogsOnFailure $ \tracer -> failAfter 30 $ do
+        [port1, port2] <- fmap fromIntegral <$> randomUnusedTCPPorts 2
+        withOuroborosNetwork tracer (Host lo port1) [Host lo port2] (const @_ @Integer $ pure ()) $ \hn1 ->
+          withOuroborosNetwork @Integer tracer (Host lo port2) [Host lo port1] (atomically . writeTQueue received) $ \_ -> do
+            withNodeBroadcastingForever hn1 1 $
+              atomically (readTQueue received) `shouldReturn` 1
+
+    it "broadcasts messages between 3 connected peers" $ do
+      node1received <- atomically newTQueue
+      node2received <- atomically newTQueue
+      node3received <- atomically newTQueue
+      showLogsOnFailure $ \tracer -> failAfter 30 $ do
+        [port1, port2, port3] <- fmap fromIntegral <$> randomUnusedTCPPorts 3
+        withOuroborosNetwork @Integer tracer (Host lo port1) [Host lo port2, Host lo port3] (atomically . writeTQueue node1received) $ \hn1 ->
+          withOuroborosNetwork tracer (Host lo port2) [Host lo port1, Host lo port3] (atomically . writeTQueue node2received) $ \hn2 -> do
+            withOuroborosNetwork tracer (Host lo port3) [Host lo port1, Host lo port2] (atomically . writeTQueue node3received) $ \hn3 -> do
+              withNodesBroadcastingForever [(hn1, 1), (hn2, 2), (hn3, 3)] $
+                assertAllnodesReceivedMessagesFromAllOtherNodes [(node1received, 1), (node2received, 2), (node3received, 3)]
 
   describe "Serialisation" $ do
-    it "can roundtrip CBOR encoding/decoding of Hydra Message" $ \_ -> property $ prop_canRoundtripCBOREncoding @(Message SimpleTx)
+    it "can roundtrip CBOR encoding/decoding of Hydra Message" $ property $ prop_canRoundtripCBOREncoding @(Message SimpleTx)
+    roundtripAndGoldenSpecs (Proxy @(Message SimpleTx))
 
 withNodeBroadcastingForever :: Network IO Integer -> Integer -> IO b -> IO b
 withNodeBroadcastingForever node value continuation = withNodesBroadcastingForever [(node, value)] continuation
