+++ hydra-test-utils/hydra-test-utils.cabal
@@ -58,12 +58,12 @@ library
     , network
     , plutus-ledger-api
     , plutus-tx
+    , port-utils
     , process
     , QuickCheck
     , relude
     , strict-containers
     , temporary
-    , warp
 
   ghc-options:      -Wall -Werror -Wcompat -Wunused-packages
   default-language: Haskell2010
