+++ hydra-test-utils/src/Test/Network/Ports.hs
@@ -6,12 +6,11 @@ module Test.Network.Ports where
 
 import Hydra.Prelude
 
-import Control.Monad.Class.MonadSTM (writeTVar)
 import Network.Socket (
   PortNumber,
   close',
  )
-import Network.Wai.Handler.Warp (openFreePort)
+import Network.Socket.Free (openFreePort)
 
 -- | Find a TCPv4 port which is likely to be free for listening on
 -- @localhost@. This binds a socket, receives an OS-assigned port, then closes
@@ -21,35 +20,22 @@ import Network.Wai.Handler.Warp (openFreePort)
 -- port returned by 'getRandomPort' before this process does.
 --
 -- Do not use this unless you have no other option.
-getRandomPort :: TVar IO [Int] -> IO PortNumber
-getRandomPort tvar = do
+getRandomPort :: IO PortNumber
+getRandomPort = do
   (port, sock) <- openFreePort
   liftIO $ close' sock
-  mport <- atomically $ recordFreePort tvar port
-  case mport of
-    Nothing -> threadDelay 1 >> getRandomPort tvar
-    Just p -> return $ fromIntegral p
+  return $ fromIntegral port
 
 -- | Find a free TCPv4 port and pass it to the given 'action'.
 --
 -- NOTE: Should be used only for testing, see 'getRandomPort' for limitations.
-withFreePort :: TVar IO [Int] -> (PortNumber -> IO a) -> IO a
-withFreePort tvar action = getRandomPort tvar >>= action
+withFreePort :: (PortNumber -> IO a) -> IO a
+withFreePort action = getRandomPort >>= action
 
 -- | Find the specified number of free ports.
 --
 -- NOTE: Should be used only for testing, see 'getRandomPort' for limitations.
-randomUnusedTCPPorts :: TVar IO [Int] -> Int -> IO [Int]
-randomUnusedTCPPorts tvar count =
+randomUnusedTCPPorts :: Int -> IO [Int]
+randomUnusedTCPPorts count =
   fmap fromIntegral
-    <$> replicateM count (withFreePort tvar (\port -> return port))
-
--- | Internal function to maybe obtain the free port.
-recordFreePort :: TVar IO [Int] -> Int -> STM IO (Maybe Int)
-recordFreePort tvar port = do
-  freePorts <- readTVar tvar
-  if port `elem` freePorts
-    then return Nothing
-    else do
-      writeTVar tvar (port : freePorts)
-      return $ Just port
+    <$> replicateM count (withFreePort (\port -> return port))
